requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
APScheduler==3.10.4
pytz==2023.3
qrcode[pil]==7.4.2
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.3
gunicorn==21.2.0
Pillow==10.1.0